+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/PlaylistController.java
@@ -3,9 +3,8 @@ package com.ciandt.summit.bootcamp2022.adapter.controller;
 import com.ciandt.summit.bootcamp2022.adapter.controller.dto.DeletedMusicFromPlaylistDto;
 import com.ciandt.summit.bootcamp2022.adapter.controller.dto.MusicInformationDto;
 import com.ciandt.summit.bootcamp2022.adapter.controller.dto.ResponseWrapper;
-import com.ciandt.summit.bootcamp2022.entity.Music;
-import com.ciandt.summit.bootcamp2022.entity.PlaylistMusicsRepository;
-import com.ciandt.summit.bootcamp2022.entity.PlaylistRepository;
+import com.ciandt.summit.bootcamp2022.entity.music.Music;
+import com.ciandt.summit.bootcamp2022.entity.playlist.PlaylistRepository;
 import com.ciandt.summit.bootcamp2022.exceptions.AuthorizedHandler;
 import com.ciandt.summit.bootcamp2022.http.TokenAuthorizedClientUtils;
 import com.ciandt.summit.bootcamp2022.repository.PlaylistMusicasImpl;
@@ -15,7 +14,6 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.apache.logging.log4j.message.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
