+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/dto/ArtistDto.java
@@ -1,6 +1,6 @@
 package com.ciandt.summit.bootcamp2022.adapter.controller.dto;
 
-import com.ciandt.summit.bootcamp2022.infra.entity.ArtistEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.artist.ArtistEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
