+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/dto/MusicDto.java
@@ -1,6 +1,6 @@
 package com.ciandt.summit.bootcamp2022.adapter.controller.dto;
 
-import com.ciandt.summit.bootcamp2022.infra.entity.MusicEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.music.MusicEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
