+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/dto/MusicInformationDto.java
@@ -1,12 +1,11 @@
 package com.ciandt.summit.bootcamp2022.adapter.controller.dto;
 
-import com.ciandt.summit.bootcamp2022.entity.Music;
+import com.ciandt.summit.bootcamp2022.entity.music.Music;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
