--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/entity/PlaylistMusicsPK.java
@@ -1,17 +1,0 @@
-package com.ciandt.summit.bootcamp2022.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-public class PlaylistMusicsPK {
-
-    private String playlistId;
-
-    private String musicId;
-}
