+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/entity/music/MusicRepository.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.entity;
+package com.ciandt.summit.bootcamp2022.entity.music;
 
 public interface MusicRepository {
