+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/entity/playlist/PlaylistMusicsRepository.java
@@ -1,4 +1,6 @@
-package com.ciandt.summit.bootcamp2022.entity;
+package com.ciandt.summit.bootcamp2022.entity.playlist;
+
+import com.ciandt.summit.bootcamp2022.entity.music.Music;
 
 public interface PlaylistMusicsRepository {
