+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/entity/playlist/PlaylistRepository.java
@@ -1,4 +1,6 @@
-package com.ciandt.summit.bootcamp2022.entity;
+package com.ciandt.summit.bootcamp2022.entity.playlist;
+
+import com.ciandt.summit.bootcamp2022.entity.music.Music;
 
 import java.util.List;
