+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/infra/entity/artist/ArtistEntity.java
@@ -1,7 +1,7 @@
-package com.ciandt.summit.bootcamp2022.infra.entity;
+package com.ciandt.summit.bootcamp2022.infra.entity.artist;
 
 
-import com.ciandt.summit.bootcamp2022.entity.Artist;
+import com.ciandt.summit.bootcamp2022.entity.artist.Artist;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
