+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/infra/entity/playlist/PlaylistEntity.java
@@ -1,7 +1,8 @@
-package com.ciandt.summit.bootcamp2022.infra.entity;
+package com.ciandt.summit.bootcamp2022.infra.entity.playlist;
 
 
-import com.ciandt.summit.bootcamp2022.entity.Playlist;
+import com.ciandt.summit.bootcamp2022.entity.playlist.Playlist;
+import com.ciandt.summit.bootcamp2022.infra.entity.music.MusicEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
