+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/repository/ArtistRepositoryWithJpa.java
@@ -1,6 +1,6 @@
 package com.ciandt.summit.bootcamp2022.repository;
 
-import com.ciandt.summit.bootcamp2022.infra.entity.ArtistEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.artist.ArtistEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
