+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/repository/PlaylistMusicsWithJpa.java
@@ -1,7 +1,7 @@
 package com.ciandt.summit.bootcamp2022.repository;
 
-import com.ciandt.summit.bootcamp2022.infra.entity.PlaylistMusicas;
-import com.ciandt.summit.bootcamp2022.infra.entity.PlaylistMusicsPKEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.playlist.PlaylistMusicas;
+import com.ciandt.summit.bootcamp2022.infra.entity.playlist.PlaylistMusicsPKEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
