+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/repository/PlaylistRepositoryWithJpa.java
@@ -1,9 +1,7 @@
 package com.ciandt.summit.bootcamp2022.repository;
 
-import com.ciandt.summit.bootcamp2022.entity.Music;
-import com.ciandt.summit.bootcamp2022.infra.entity.PlaylistEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.playlist.PlaylistEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 public interface PlaylistRepositoryWithJpa extends JpaRepository<PlaylistEntity, String>{
