+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/usecase/MusicService.java
@@ -1,6 +1,6 @@
 package com.ciandt.summit.bootcamp2022.usecase;
 
-import com.ciandt.summit.bootcamp2022.infra.entity.MusicEntity;
+import com.ciandt.summit.bootcamp2022.infra.entity.music.MusicEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
